requirements.txt
blinker==1.9.0
-e git+ssh://git@github.com/nannac/christophernanna-me.git@29de3c809feccb1a1c1ad9ea67fc8a0798689aeb#egg=christophernanna_me
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
python-dotenv==1.2.1
python-frontmatter==1.1.0
PyYAML==6.0.3
Werkzeug==3.1.5
